requirements.txt
beautifulsoup4
requests
python-dotenv
supabase
# sentence-transformers
numpy==1.26.4
google-genai
pandas
lxml
# pytrends
# schedule
python-dotenv